// frontend/src/backend/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
      colleges: {
        Row: {
          college_name: string;
          id: number;
        };
        Insert: {
          college_name: string;
          id?: number;
        };
        Update: {
          college_name?: string;
          id?: number;
        };
        Relationships: [];
      };
// ...
      subjects: {
        Row: {
          id: number;
          subject_name: string;
        };
        Insert: {
          id?: number;
          subject_name: string;
        };
        Update: {
          id?: number;
          subject_name?: string;
        };
        Relationships: [];
      };
      users: {
        Row: {
          created_at: string;
          email: string;
          full_name: string | null;
          id: string;
          password_hash: string;
          role_id: number;
        };
        Insert: {
          created_at?: string;
          email: string;
          full_name?: string | null;
          id?: string;
          password_hash: string;
          role_id: number;
        };
        Update: {
          created_at?: string;
          email?: string;
          full_name?: string | null;
          id?: string;
          password_hash?: string;
          role_id?: number;
        };
        Relationships: [
          {
            foreignKeyName: "users_role_id_fkey";
            columns: ["role_id"];
            isOneToOne: false;
            referencedRelation: "roles";
            referencedColumns: ["id"];
          },
        ];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      generate_jwt_token: {
        Args: {
          user_id: string;
        };
        Returns: string;
      };
      get_student: {
        Args: Record<PropertyKey, never>;
        Returns: {
          id: number;
          generation: number;
          khmer_name: string;
          english_name: string;
          date_of_birth: string;
          sex: string;
          nationality: string;
          address: string;
          phone_number: string;
          facebook_or_email: string;
          father_name: string;
          father_nationality: string;
          father_occupation: string;
          father_phone_number: string;
          mother_name: string;
          mother_nationality: string;
          mother_occupation: string;
          mother_phone_number: string;
          session: string;
          session_id: number;
          subject: string;
          subject_id: number;
          college: string;
          college_id: number;
          grade: string;
          grade_id: number;
          family_situation: string;
          family_situation_id: number;
          student_job: string;
          student_job_id: number;
        }[];
      };
// ...
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};
type PublicSchema = Database[Extract<keyof Database, "public">];

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R;
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R;
      }
      ? R
      : never
    : never;
// ...
export type Functions<
  PublicFunctionNameOrOptions extends
    | keyof PublicSchema["Functions"]
    | { schema: keyof Database },
  FunctionName extends PublicFunctionNameOrOptions extends {
    schema: keyof Database;
  }
    ? keyof Database[PublicFunctionNameOrOptions["schema"]]["Functions"]
    : never = never,
> = PublicFunctionNameOrOptions extends { schema: keyof Database }
  ? Database[PublicFunctionNameOrOptions["schema"]]["Functions"][FunctionName] extends {
      Returns: infer R;
    }
    ? R
    : never
  : PublicFunctionNameOrOptions extends keyof PublicSchema["Functions"]
    ? PublicSchema["Functions"][PublicFunctionNameOrOptions] extends {
        Returns: infer R;
      }
      ? R
      : never
    : never;